Implemented [UpgraderState] that is used internally to replace evaluation ready. Removed Appcast configuration so that an Appcast [UpgraderStore] can be used.
