+++ lib/src/upgrade_alert.dart
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021-2023 Larry Aasen. All rights reserved.
+ * Copyright (c) 2021-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import 'upgrade_messages.dart';
+import 'upgrade_state.dart';
 import 'upgrader.dart';
 
 /// There are two different dialog styles: Cupertino and Material
@@ -99,29 +100,30 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   /// Describes the part of the user interface represented by this widget.
   @override
   Widget build(BuildContext context) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: build UpgradeAlert');
     }
 
     return StreamBuilder(
-      initialData: widget.upgrader.evaluationReady,
-      stream: widget.upgrader.evaluationStream,
-      builder:
-          (BuildContext context, AsyncSnapshot<UpgraderEvaluateNeed> snapshot) {
+      initialData: widget.upgrader.state,
+      stream: widget.upgrader.stateStream,
+      builder: (BuildContext context, AsyncSnapshot<UpgraderState> snapshot) {
         if ((snapshot.connectionState == ConnectionState.waiting ||
                 snapshot.connectionState == ConnectionState.active) &&
-            snapshot.data != null &&
-            snapshot.data!) {
-          if (widget.upgrader.debugLogging) {
-            print("upgrader: need to evaluate version");
-          }
+            snapshot.data != null) {
+          final upgraderState = snapshot.data!;
+          if (upgraderState.versionInfo != null) {
+            if (widget.upgrader.state.debugLogging) {
+              print("upgrader: need to evaluate version");
+            }
 
-          if (!displayed) {
-            final checkContext = widget.navigatorKey != null &&
-                    widget.navigatorKey!.currentContext != null
-                ? widget.navigatorKey!.currentContext!
-                : context;
-            checkVersion(context: checkContext);
+            if (!displayed) {
+              final checkContext = widget.navigatorKey != null &&
+                      widget.navigatorKey!.currentContext != null
+                  ? widget.navigatorKey!.currentContext!
+                  : context;
+              checkVersion(context: checkContext);
+            }
           }
         }
         return widget.child ?? const SizedBox.shrink();
@@ -133,7 +135,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   /// Only called by [UpgradeAlert] and not used by [UpgradeCard].
   void checkVersion({required BuildContext context}) {
     final shouldDisplay = widget.upgrader.shouldDisplayUpgrade();
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: shouldDisplayReleaseNotes: $shouldDisplayReleaseNotes');
     }
     if (shouldDisplay) {
@@ -156,7 +158,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserIgnored(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: ignore');
     }
 
@@ -173,7 +175,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserLater(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: later');
     }
 
@@ -186,7 +188,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserUpdated(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: update now');
     }
 
@@ -221,7 +223,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
     required bool canDismissDialog,
     required UpgraderMessages messages,
   }) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: showTheDialog title: $title');
       print('upgrader: showTheDialog message: $message');
       print('upgrader: showTheDialog releaseNotes: $releaseNotes');
@@ -253,12 +255,12 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   /// is false. Also called when the back button is pressed. Return true for
   /// the screen to be popped. Defaults to false.
   bool onWillPop() {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: onWillPop called');
     }
     if (widget.shouldPopScope != null) {
       final should = widget.shouldPopScope!();
-      if (widget.upgrader.debugLogging) {
+      if (widget.upgrader.state.debugLogging) {
         print('upgrader: shouldPopScope=$should');
       }
       return should;
