+++ lib/src/upgrade_card.dart
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021-2023 Larry Aasen. All rights reserved.
+ * Copyright (c) 2021-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/material.dart';
 
 import 'alert_style_widget.dart';
 import 'upgrade_messages.dart';
+import 'upgrade_state.dart';
 import 'upgrader.dart';
 
 /// A widget to display the upgrade card.
@@ -78,25 +79,26 @@ class UpgradeCardState extends State<UpgradeCard> {
   /// Describes the part of the user interface represented by this widget.
   @override
   Widget build(BuildContext context) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: build UpgradeCard');
     }
 
     return StreamBuilder(
-        initialData: widget.upgrader.evaluationReady,
-        stream: widget.upgrader.evaluationStream,
-        builder: (BuildContext context,
-            AsyncSnapshot<UpgraderEvaluateNeed> snapshot) {
+        initialData: widget.upgrader.state,
+        stream: widget.upgrader.stateStream,
+        builder: (BuildContext context, AsyncSnapshot<UpgraderState> snapshot) {
           if ((snapshot.connectionState == ConnectionState.waiting ||
                   snapshot.connectionState == ConnectionState.active) &&
-              snapshot.data != null &&
-              snapshot.data!) {
-            if (widget.upgrader.shouldDisplayUpgrade()) {
-              return buildUpgradeCard(
-                  context, const Key('upgrader_alert_card'));
-            } else {
-              if (widget.upgrader.debugLogging) {
-                print('upgrader: UpgradeCard will not display');
+              snapshot.data != null) {
+            final upgraderState = snapshot.data!;
+            if (upgraderState.versionInfo != null) {
+              if (widget.upgrader.shouldDisplayUpgrade()) {
+                return buildUpgradeCard(
+                    context, const Key('upgrader_alert_card'));
+              } else {
+                if (widget.upgrader.state.debugLogging) {
+                  print('upgrader: UpgradeCard will not display');
+                }
               }
             }
           }
@@ -111,7 +113,7 @@ class UpgradeCardState extends State<UpgradeCard> {
     final message = widget.upgrader.body(appMessages);
     final releaseNotes = widget.upgrader.releaseNotes;
 
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: UpgradeCard: will display');
       print('upgrader: UpgradeCard: showDialog title: $title');
       print('upgrader: UpgradeCard: showDialog message: $message');
@@ -208,7 +210,7 @@ class UpgradeCardState extends State<UpgradeCard> {
       (widget.upgrader.releaseNotes?.isNotEmpty ?? false);
 
   void onUserIgnored() {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: ignore');
     }
 
@@ -223,7 +225,7 @@ class UpgradeCardState extends State<UpgradeCard> {
   }
 
   void onUserLater() {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: later');
     }
 
@@ -234,7 +236,7 @@ class UpgradeCardState extends State<UpgradeCard> {
   }
 
   void onUserUpdated() {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: update now');
     }
