+++ test/fake_appcast.dart
@@ -5,7 +5,6 @@ import 'dart:io';
 
 import 'package:mockito/mockito.dart';
 import 'package:upgrader/src/appcast.dart';
-import 'package:upgrader/src/upgrader.dart';
 
 import 'appcast_test.dart';
 
@@ -55,16 +54,16 @@ class FakeAppcast extends Fake implements TestAppcast {
     return [AppcastItem()];
   }
 
-  AppcastConfiguration config =
-      AppcastConfiguration(url: 'http://some.fakewebsite.com', supportedOS: [
-    'android',
-    'fuchsia',
-    'ios',
-    'linux',
-    'macos',
-    'web',
-    'windows',
-  ]);
+  // AppcastConfiguration config =
+  //     AppcastConfiguration(url: 'http://some.fakewebsite.com', supportedOS: [
+  //   'android',
+  //   'fuchsia',
+  //   'ios',
+  //   'linux',
+  //   'macos',
+  //   'web',
+  //   'windows',
+  // ]);
 
   @override
   List<AppcastItem>? items = [];
